ビルドエラーの修正 ビルド警告の削減

+++ src/FPS.hpp
@@ -14,7 +14,7 @@ class GameWindow;
 
 class FPS {
 private:
-    const int MAX_FPS_VECTOR_COUNT = 30;
+    const size_t MAX_FPS_VECTOR_COUNT = 30;
     std::queue<uint64_t> ticks_queue;
 public:
     void Update();
